requirements.txt
Flask
python-telegram-bot==20.7
web3
solana
python-dotenv
requests
